feat/renamed files to foo-bar.js format
